requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
pyarrow>=16.0.0
pandera>=0.18.0
apache-airflow>=2.9.0
pytest>=8.0.0
